refactor(settings): AppTab extrahiert (H4 Schritt 4)
